requirements.txt
streamlit
sqlalchemy
pillow
opencv-contrib-python
tensorflow
pandas
numpy
seaborn
matplotlib
fer